+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Qblox backend - Let the compiler raise an error when the waveforms specified in the schedule are too large to be uploaded to a sequencer.
 - Qblox backend - Forbid repeated acquisition index in schedule (!655, partially revert !542)
 - Qblox backend - Rename the `MAX_SAMPLE_SIZE_ACQUISITIONS` constant to `MAX_SAMPLE_SIZE_SCOPE_ACQUISITIONS`, and modify the docstring to clarify that this constant only refers to scope trace acquisitions (!649).
+- Qblox backend - Forbid repeated acquisition index in schedule (!655, !657, partially revert !542)
 
 ## 0.12.3 (2023-03-09)